requirements.txt
tensorflow==2.0.0
ipykernel==5.3.0
scipy==1.3.1
pandas==1.0.1
matplotlib==3.0.3
argparse
sklearn
python-binance
colorlog